Adding some OpenCV stuff

+++ pumpkin_py_pkg/pumpkin_py_pkg/pi2_ui.py
@@ -9,7 +9,10 @@ import time
 import subprocess
 import psutil
 from tkinter import Canvas, Frame, IntVar, messagebox
-from example_interfaces.srv import SetBool
+import cv2
+import numpy as np
+from std_srvs.srv import SetBool
+from PIL import Image, ImageTk
 
 #------------------------------------------------------------
 # class Pi2UI
@@ -50,6 +53,14 @@ class Pi2UI(tk.Tk):
         rightFrame.pack_propagate(0)
         rightFrame.grid(row=0, column=2, pady=(10,10))
 
+        self.faceDetectionButton = tk.Button(leftFrame, text = "Face Detection", takefocus=0, width=100, 
+                command=self.faceDetectionButtonPressed)
+        self.faceDetectionButton.pack(anchor="center", padx=(10,10), pady=(10,10))
+
+        self.laneDetectionButton = tk.Button(leftFrame, text = "Lane Detection", takefocus=0, width=100, 
+                command=self.laneDetectionButtonPressed)
+        self.laneDetectionButton.pack(anchor="center", padx=(10,10), pady=(10,10))
+
         quitButton = tk.Button(leftFrame, text = "Quit", takefocus=0, width=100, command=self.quitButtonCallback)
         quitButton.pack(anchor="center", padx=(10,10), pady=(10,10))
 
@@ -57,12 +68,32 @@ class Pi2UI(tk.Tk):
 
         self.statusLabel.config(text="Status: Waiting for Initialization Command...")
 
+        self.startedNavigation = False
+        self.startedSLAM = False
+        self.startedLaneDetection = False
+        self.startedFaceDetection = False
+
+        self.videoCapture = cv2.VideoCapture(0)
+        self.laneDetectionImage = self.videoCapture.read()
+        self.faceDetectionImage = self.videoCapture.read()
+
+        # setting the width and height of the video window
+        self.videoCapture.set(3, 640)
+        self.videoCapture.set(4, 480)
+
+        self.cascade_classifier = cv2.CascadeClassifier('/home/ubuntu/ros2_ws/src/pumpkin_py_pkg/pumpkin_py_pkg/haarcascade_frontalface_alt.xml')
+
+        self.openCVCanvas = Canvas(self, width=600, height=600, takefocus=0)  
+        self.openCVCanvas.grid(row=0, column=1, pady=(10,0))
+        
     def startNavigation(self):
+        self.startedNavigation = True
         #self.navigationProcess = subprocess.Popen(["ros2", "launch", "nav2_bringup", "bringup_launch.py", "use_sim_time:=false", "params_file:=/home/ubuntu/ros2_ws/src/pumpkin_bot_bringup/config/nav2_params.yaml", "map:=/home/ubuntu/bonus_room.yaml"])
         self.navigationProcess = subprocess.Popen(["ros2", "launch", "nav2_bringup", "navigation_launch.py", "use_sim_time:=false", "params_file:=/home/ubuntu/ros2_ws/src/pumpkin_bot_bringup/config/nav2_params.yaml"])
         time.sleep(1)
 
     def startSLAM(self):
+        self.startedSLAM = True
         #self.slamProcess = subprocess.Popen(["ros2", "launch", "slam_toolbox", "offline_launch.py", "use_sim_time:=false", "params_file:=/home/ubuntu/ros2_ws/src/pumpkin_bot_bringup/config/slam_toolbox_mapper_params_offline.yaml"])
         self.slamProcess = subprocess.Popen(["ros2", "launch", "slam_toolbox", "online_async_launch.py", "use_sim_time:=false", "params_file:=/home/ubuntu/ros2_ws/src/pumpkin_bot_bringup/config/slam_toolbox_mapper_params_online_async.yaml"])
         #self.slamProcess = subprocess.Popen(["ros2", "launch", "slam_toolbox", "online_sync_launch.py", "use_sim_time:=false", "params_file:=/home/ubuntu/ros2_ws/src/pumpkin_bot_bringup/config/slam_toolbox_mapper_params_online_sync.yaml"])
@@ -74,17 +105,23 @@ class Pi2UI(tk.Tk):
 
             self.statusLabel.config(text="Status: Quitting application...")
     
-            # Try to kill the SLAM process so the node doesn't stick around
-            slamProcess = psutil.Process(self.slamProcess.pid)
-            for proc in slamProcess.children(recursive=True):
-                proc.kill()
-            slamProcess.kill()
+            if self.startedSLAM:
+                # Try to kill the SLAM process so the node doesn't stick around
+                slamProcess = psutil.Process(self.slamProcess.pid)
+                for proc in slamProcess.children(recursive=True):
+                    proc.kill()
+                slamProcess.kill()
 
             # Try to kill the Navigation process so the node doesn't stick around
-            navigationProcess = psutil.Process(self.navigationProcess.pid)
-            for proc in navigationProcess.children(recursive=True):
-                proc.kill()
-            navigationProcess.kill()
+            if self.startedNavigation:
+                navigationProcess = psutil.Process(self.navigationProcess.pid)
+                for proc in navigationProcess.children(recursive=True):
+                    proc.kill()
+                navigationProcess.kill()
+
+            if self.startedLaneDetection:
+                self.videoCapture.release()
+                cv2.destroyAllWindows()
 
             self.destroy()  
 
@@ -101,21 +138,157 @@ class Pi2UI(tk.Tk):
             self.statusLabel.config(text="Status: Quitting Application...")
     
             # Try to kill the SLAM process so the node doesn't stick around
-            slamProcess = psutil.Process(self.slamProcess.pid)
-            for proc in slamProcess.children(recursive=True):
-                proc.kill()
-            slamProcess.kill()
+            if self.startedSLAM:
+                slamProcess = psutil.Process(self.slamProcess.pid)
+                for proc in slamProcess.children(recursive=True):
+                    proc.kill()
+                slamProcess.kill()
 
             # Try to kill the Navigation process so the node doesn't stick around
-            navigationProcess = psutil.Process(self.navigationProcess.pid)
-            for proc in navigationProcess.children(recursive=True):
-                proc.kill()
-            navigationProcess.kill()
+            if self.startedNavigation:
+                navigationProcess = psutil.Process(self.navigationProcess.pid)
+                for proc in navigationProcess.children(recursive=True):
+                    proc.kill()
+                navigationProcess.kill()
 
             self.destroy()      
 
     def startMainLoop(self):
         self.mainloop() 
+
+
+    def draw_the_lines(self, image, lines):
+        # create a distinct image for the lines [0,255] - all 0 values means black image
+        lines_image = np.zeros((image.shape[0], image.shape[1], 3), dtype=np.uint8)
+
+        # there are (x,y) for the starting and end points of the lines
+        if lines is not None:
+            for line in lines:
+                for x1, y1, x2, y2 in line:
+                    cv2.line(lines_image, (x1, y1), (x2, y2), (255, 0, 0), thickness=3)
+
+        # finally we have to merge the image with the lines
+        image_with_lines = cv2.addWeighted(image, 0.8, lines_image, 1, 0.0)
+
+        return image_with_lines
+
+
+    def region_of_interest(self, image, region_points):
+        # we are going to replace pixels with 0 (black) - the regions we are not interested
+        mask = np.zeros_like(image)
+        # the region that we are interested in is the lower triangle - 255 white pixels
+        cv2.fillPoly(mask, region_points, 255)
+        # we have to use the mask: we want to keep the regions of the original image where  
+        # the mask has white colored pixels
+        masked_image = cv2.bitwise_and(image, mask)
+        return masked_image
+
+
+    def get_detected_lanes(self, image):
+
+        _, image = self.videoCapture.read()
+
+        (height, width) = (image.shape[0], image.shape[1])
+
+        # we have to turn the image into grayscale
+        gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
+
+        # edge detection kernel (Canny's algorithm)
+        canny_image = cv2.Canny(gray_image, 100, 120)
+
+        # we are interested in the "lower region" of the image (there are the driving lanes)
+        region_of_interest_vertices = [
+            (0, height),
+            (width / 2, height * 0.65),
+            (width, height)
+        ]
+
+        # we can get rid of the un-relevant part of the image
+        # we just keep the lower triangle region
+        cropped_image = self.region_of_interest(
+            canny_image, np.array([region_of_interest_vertices], np.int32))
+
+        # use the line detection algorithm (radians instead of degrees 1 degree = pi / 180)
+        lines = cv2.HoughLinesP(cropped_image, rho=2, theta=np.pi / 180, threshold=50, lines=np.array([]),
+                            minLineLength=40, maxLineGap=150)
+
+        # draw the lines on the image
+        image_with_lines = self.draw_the_lines(image, lines)
+
+        return image_with_lines
+    
+ 
+    def startLaneDetection(self):
+ 
+        self.laneDetectionImage = self.get_detected_lanes(self.laneDetectionImage)
+
+        self.laneDetectionImage = cv2.cvtColor(self.laneDetectionImage, cv2.COLOR_BGR2RGB)
+        self.laneDetectionImage = Image.fromarray(self.laneDetectionImage) # to PIL format
+        self.laneDetectionImage = ImageTk.PhotoImage(self.laneDetectionImage) # to ImageTk format
+        # Update image
+        self.openCVCanvas.create_image(0, 0, anchor=tk.NW, image=self.laneDetectionImage)
+        # Repeat every 'interval' ms
+        self.laneDetectionCancelId = self.after(10, self.startLaneDetection)
+
+
+    def laneDetectionButtonPressed(self):
+        if self.startedLaneDetection:
+            self.startedLaneDetection = False
+            self.laneDetectionButton['text'] = 'Lane Detection'
+            self.openCVCanvas.delete("all")
+            self.after_cancel(self.laneDetectionCancelId)
+        elif not self.startedFaceDetection:
+            self.openCVCanvas.delete("all")
+            self.startedLaneDetection = True
+            self.laneDetectionButton['text'] = 'Lane Detection*'
+            self.startLaneDetection()
+
+
+    def get_detected_faces(self, image):
+        
+        _, image = self.videoCapture.read()
+
+        (height, width) = (image.shape[0], image.shape[1])
+
+        # we have to turn the image into grayscale
+        gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
+
+        # using face detection algorithm with the trained classifier
+        detected_faces = self.cascade_classifier.detectMultiScale(gray_image, scaleFactor=1.2, minNeighbors=5, minSize=(30, 30),
+                                 flags=cv2.CASCADE_SCALE_IMAGE)
+
+        # draw the faces (rectangle) in every video frame
+        for (x, y, width, height) in detected_faces:
+            cv2.rectangle(image, (x, y), (x + width, y + height), (0, 0, 255), 10)
+
+        return image
+
+
+    def startFaceDetection(self):
+
+        self.faceDetectionImage = self.get_detected_faces(self.faceDetectionImage)
+
+        self.faceDetectionImage = cv2.cvtColor(self.faceDetectionImage, cv2.COLOR_BGR2RGB)
+        self.faceDetectionImage = Image.fromarray(self.faceDetectionImage) # to PIL format
+        self.faceDetectionImage = ImageTk.PhotoImage(self.faceDetectionImage) # to ImageTk format
+        # Update image
+        self.openCVCanvas.create_image(0, 0, anchor=tk.NW, image=self.faceDetectionImage)
+        # Repeat every 'interval' ms
+        self.faceDetectionCancelId = self.after(10, self.startFaceDetection)
+
+
+    def faceDetectionButtonPressed(self):
+        if self.startedFaceDetection:
+            self.startedFaceDetection = False
+            self.faceDetectionButton['text'] = 'Face Detection'
+            self.openCVCanvas.delete("all")
+            self.after_cancel(self.faceDetectionCancelId)
+        elif not self.startedLaneDetection:
+            self.openCVCanvas.delete("all")
+            self.startedFaceDetection = True
+            self.faceDetectionButton['text'] = 'Face Detection*'
+            self.startFaceDetection()
+
 
 #------------------------------------------------------------
 # class Pi2UserInterfaceNode
